Add jsdocs and refactor bubble to use new metrics
