+++ core/bubble.js
@@ -241,8 +241,8 @@ Blockly.Bubble.prototype.createDom_ = function(content, hasResize) {
     [...content goes here...]
   </g>
   */
-  this.bubbleGroup_ = Blockly.utils.dom.createSvgElement(
-      Blockly.utils.Svg.G, {}, null);
+  this.bubbleGroup_ =
+      Blockly.utils.dom.createSvgElement(Blockly.utils.Svg.G, {}, null);
   var filter = {
     'filter': 'url(#' +
         this.workspace_.getRenderer().getConstants().embossFilterId + ')'
@@ -458,14 +458,10 @@ Blockly.Bubble.prototype.setAnchorLocation = function(xy) {
  */
 Blockly.Bubble.prototype.layoutBubble_ = function() {
   // Get the metrics in workspace units.
-  var metrics = this.workspace_.getMetrics();
-  metrics.viewLeft /= this.workspace_.scale;
-  metrics.viewWidth /= this.workspace_.scale;
-  metrics.viewTop /= this.workspace_.scale;
-  metrics.viewHeight /= this.workspace_.scale;
+  var viewMetrics = this.workspace_.metricsManager.getViewMetrics(true);
 
-  var optimalLeft = this.getOptimalRelativeLeft_(metrics);
-  var optimalTop = this.getOptimalRelativeTop_(metrics);
+  var optimalLeft = this.getOptimalRelativeLeft_(viewMetrics);
+  var optimalTop = this.getOptimalRelativeTop_(viewMetrics);
   var bbox = this.shape_.getBBox();
 
   var topPosition = {
@@ -480,10 +476,10 @@ Blockly.Bubble.prototype.layoutBubble_ = function() {
   var closerPosition = bbox.width < bbox.height ? endPosition : bottomPosition;
   var fartherPosition = bbox.width < bbox.height ? bottomPosition : endPosition;
 
-  var topPositionOverlap = this.getOverlap_(topPosition, metrics);
-  var startPositionOverlap = this.getOverlap_(startPosition, metrics);
-  var closerPositionOverlap = this.getOverlap_(closerPosition, metrics);
-  var fartherPositionOverlap = this.getOverlap_(fartherPosition, metrics);
+  var topPositionOverlap = this.getOverlap_(topPosition, viewMetrics);
+  var startPositionOverlap = this.getOverlap_(startPosition, viewMetrics);
+  var closerPositionOverlap = this.getOverlap_(closerPosition, viewMetrics);
+  var fartherPositionOverlap = this.getOverlap_(fartherPosition, viewMetrics);
 
   // Set the position to whichever position shows the most of the bubble,
   // with tiebreaks going in the order: top > start > close > far.
@@ -517,12 +513,12 @@ Blockly.Bubble.prototype.layoutBubble_ = function() {
  * workspace (what percentage of the bubble is visible).
  * @param {!{x: number, y: number}} relativeMin The position of the top-left
  *     corner of the bubble relative to the anchor point.
- * @param {!Blockly.utils.Metrics} metrics The metrics of the workspace the
- *     bubble will appear in.
+ * @param {!Blockly.MetricsManager.ContainerSize} viewMetrics The metrics of the
+ *     workspace the bubble will appear in.
  * @return {number} The percentage of the bubble that is visible.
  * @private
  */
-Blockly.Bubble.prototype.getOverlap_ = function(relativeMin, metrics) {
+Blockly.Bubble.prototype.getOverlap_ = function(relativeMin, viewMetrics) {
   // The position of the top-left corner of the bubble in workspace units.
   var bubbleMin = {
     x: this.workspace_.RTL ? (this.anchorXY_.x - relativeMin.x - this.width_) :
@@ -538,11 +534,11 @@ Blockly.Bubble.prototype.getOverlap_ = function(relativeMin, metrics) {
   // calculation.
 
   // The position of the top-left corner of the workspace.
-  var workspaceMin = {x: metrics.viewLeft, y: metrics.viewTop};
+  var workspaceMin = {x: viewMetrics.left, y: viewMetrics.top};
   // The position of the bottom-right corner of the workspace.
   var workspaceMax = {
-    x: metrics.viewLeft + metrics.viewWidth,
-    y: metrics.viewTop + metrics.viewHeight
+    x: viewMetrics.left + viewMetrics.width,
+    y: viewMetrics.top + viewMetrics.height
   };
 
   var overlapWidth = Math.min(bubbleMax.x, workspaceMax.x) -
@@ -559,17 +555,17 @@ Blockly.Bubble.prototype.getOverlap_ = function(relativeMin, metrics) {
  * Calculate what the optimal horizontal position of the top-left corner of the
  * bubble is (relative to the anchor point) so that the most area of the
  * bubble is shown.
- * @param {!Blockly.utils.Metrics} metrics The metrics of the workspace the
- *     bubble will appear in.
+ * @param {!Blockly.MetricsManager.ContainerSize} viewMetrics The view metrics of the
+ *     workspace the bubble will appear in.
  * @return {number} The optimal horizontal position of the top-left corner
  *     of the bubble.
  * @private
  */
-Blockly.Bubble.prototype.getOptimalRelativeLeft_ = function(metrics) {
+Blockly.Bubble.prototype.getOptimalRelativeLeft_ = function(viewMetrics) {
   var relativeLeft = -this.width_ / 4;
 
   // No amount of sliding left or right will give us a better overlap.
-  if (this.width_ > metrics.viewWidth) {
+  if (this.width_ > viewMetrics.width) {
     return relativeLeft;
   }
 
@@ -578,16 +574,16 @@ Blockly.Bubble.prototype.getOptimalRelativeLeft_ = function(metrics) {
     var bubbleRight = this.anchorXY_.x - relativeLeft;
     var bubbleLeft = bubbleRight - this.width_;
 
-    var workspaceRight = metrics.viewLeft + metrics.viewWidth;
-    var workspaceLeft = metrics.viewLeft +
+    var workspaceRight = viewMetrics.left + viewMetrics.width;
+    var workspaceLeft = viewMetrics.left +
         // Thickness in workspace units.
         (Blockly.Scrollbar.scrollbarThickness / this.workspace_.scale);
   } else {
     var bubbleLeft = relativeLeft + this.anchorXY_.x;
     var bubbleRight = bubbleLeft + this.width_;
 
-    var workspaceLeft = metrics.viewLeft;
-    var workspaceRight = metrics.viewLeft + metrics.viewWidth -
+    var workspaceLeft = viewMetrics.left;
+    var workspaceRight = viewMetrics.left + viewMetrics.width -
         // Thickness in workspace units.
         (Blockly.Scrollbar.scrollbarThickness / this.workspace_.scale);
   }
@@ -617,24 +613,24 @@ Blockly.Bubble.prototype.getOptimalRelativeLeft_ = function(metrics) {
  * Calculate what the optimal vertical position of the top-left corner of
  * the bubble is (relative to the anchor point) so that the most area of the
  * bubble is shown.
- * @param {!Blockly.utils.Metrics} metrics The metrics of the workspace the
- *     bubble will appear in.
+ * @param {!Blockly.MetricsManager.ContainerSize} viewMetrics The view metrics of the
+ *     workspace the bubble will appear in.
  * @return {number} The optimal vertical position of the top-left corner
  *     of the bubble.
  * @private
  */
-Blockly.Bubble.prototype.getOptimalRelativeTop_ = function(metrics) {
+Blockly.Bubble.prototype.getOptimalRelativeTop_ = function(viewMetrics) {
   var relativeTop = -this.height_ / 4;
 
   // No amount of sliding up or down will give us a better overlap.
-  if (this.height_ > metrics.viewHeight) {
+  if (this.height_ > viewMetrics.height) {
     return relativeTop;
   }
 
   var bubbleTop = this.anchorXY_.y + relativeTop;
   var bubbleBottom = bubbleTop + this.height_;
-  var workspaceTop = metrics.viewTop;
-  var workspaceBottom = metrics.viewTop + metrics.viewHeight -
+  var workspaceTop = viewMetrics.top;
+  var workspaceBottom = viewMetrics.top + viewMetrics.height -
       // Thickness in workspace units.
       (Blockly.Scrollbar.scrollbarThickness / this.workspace_.scale);
 
@@ -891,8 +887,7 @@ Blockly.Bubble.prototype.setAutoLayout = function(enable) {
  */
 Blockly.Bubble.textToDom = function(text) {
   var paragraph = Blockly.utils.dom.createSvgElement(
-      Blockly.utils.Svg.TEXT,
-      {
+      Blockly.utils.Svg.TEXT, {
         'class': 'blocklyText blocklyBubbleText blocklyNoPointerEvents',
         'y': Blockly.Bubble.BORDER_WIDTH
       },
@@ -910,16 +905,18 @@ Blockly.Bubble.textToDom = function(text) {
 
 /**
  * Creates a bubble that can not be edited.
- * @param {!SVGTextElement} paragraphElement The text element for the non editable bubble.
+ * @param {!SVGTextElement} paragraphElement The text element for the non
+ *     editable bubble.
  * @param {!Blockly.BlockSvg} block The block that the bubble is attached to.
  * @param {!Blockly.utils.Coordinate} iconXY The coordinate of the icon.
  * @return {!Blockly.Bubble} The non editable bubble.
  * @package
  */
-Blockly.Bubble.createNonEditableBubble = function(paragraphElement, block, iconXY) {
+Blockly.Bubble.createNonEditableBubble = function(
+    paragraphElement, block, iconXY) {
   var bubble = new Blockly.Bubble(
-      /** @type {!Blockly.WorkspaceSvg} */ (block.workspace),
-      paragraphElement, block.pathObject.svgPath,
+      /** @type {!Blockly.WorkspaceSvg} */ (block.workspace), paragraphElement,
+      block.pathObject.svgPath,
       /** @type {!Blockly.utils.Coordinate} */ (iconXY), null, null);
   // Expose this bubble's block's ID on its top-level SVG group.
   bubble.setSvgId(block.id);
@@ -927,9 +924,8 @@ Blockly.Bubble.createNonEditableBubble = function(paragraphElement, block, iconX
     // Right-align the paragraph.
     // This cannot be done until the bubble is rendered on screen.
     var maxWidth = paragraphElement.getBBox().width;
-    for (var i = 0, textElement;
-      (textElement = paragraphElement.childNodes[i]); i++) {
-
+    for (var i = 0, textElement; (textElement = paragraphElement.childNodes[i]);
+         i++) {
       textElement.setAttribute('text-anchor', 'end');
       textElement.setAttribute('x', maxWidth + Blockly.Bubble.BORDER_WIDTH);
     }
